refactor clob client
